@@ -1,6 +1,6 @@
 {
   "name": "localstorage-fifo",
-  "version": "2.0.2-pre",
+  "version": "3.0.0",
   "description": "JavaScript library for interacting with localStorage safely.",
   "author": "Matthew Callis <matthew.callis@gmail.com>",
   "contributors": [
@@ -9,7 +9,6 @@
       "email": "matthew.callis@gmail.com"
     }
   ],
-  "main": "lib/fifo.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/MatthewCallis/fifo"
@@ -21,52 +20,58 @@
     "localStorage",
     "fifo"
   ],
+  "files": [
+    "lib/*",
+    "src/*",
+    "types/*"
+  ],
+  "main": "lib/fifo.js",
+  "module": "./src/fifo.js",
+  "type": "module",
+  "types": "types/index.d.ts",
+  "typings": "types/index.d.ts",
+  "sideEffects": false,
   "dependencies": {},
   "devDependencies": {
-    "ava": "^0.25.0",
-    "babel-cli": "^6.26.0",
-    "babel-plugin-espower": "^2.4.0",
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-istanbul": "^4.1.5",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-es2015-rollup": "^3.0.0",
-    "babel-preset-es2016": "^6.24.1",
-    "babel-register": "^6.26.0",
-    "browser-env": "^3.2.5",
-    "codeclimate-test-reporter": "^0.5.0",
-    "coveralls": "^3.0.0",
-    "eslint": "^4.18.2",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-ava": "^4.5.1",
-    "eslint-plugin-import": "^2.9.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-no-inferred-method-name": "^1.0.2",
-    "eslint-plugin-react": "^7.7.0",
-    "eslint-plugin-xss": "^0.1.9",
-    "leakage": "^0.3.0",
-    "npm-bump": "^0.0.23",
-    "nyc": "^11.4.1",
-    "precommit-hook-eslint": "^3.0.0",
-    "rollup": "^0.56.5",
-    "rollup-plugin-babel": "^3.0.3",
-    "rollup-plugin-eslint": "^4.0.0",
-    "rollup-plugin-json": "^2.3.0",
-    "uglify-js": "^3.3.13"
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-commonjs": "^21.0.1",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@typescript-eslint/parser": "^5.12.1",
+    "ava": "^4.0.1",
+    "browser-env": "^3.3.0",
+    "c8": "^7.11.0",
+    "eslint": "^8.9.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-ava": "^13.2.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jest": "^26.1.1",
+    "eslint-plugin-jsdoc": "^37.9.4",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-optimize-regex": "^1.2.1",
+    "eslint-plugin-ramda": "^2.5.1",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-security": "^1.4.0",
+    "eslint-plugin-xss": "^0.1.11",
+    "jsdoc": "^3.6.10",
+    "jsdoc-tsimport-plugin": "^1.0.5",
+    "rollup": "^2.68.0",
+    "rollup-plugin-cleanup": "^3.2.1",
+    "typescript": "^4.5.5"
+  },
+  "engines": {
+    "node": "^16"
   },
   "preferGlobal": false,
   "private": false,
   "license": "MIT",
-  "jam": {
-    "dependencies": {}
-  },
   "scripts": {
-    "lint": "eslint --quiet --color src/",
-    "local-web-server": "ruby -run -ehttpd . -p12345",
+    "lint": "eslint src/",
     "make": "node rollup.config.js",
-    "report": "nyc report --reporter=html",
     "test-watch": "npm test -- --watch",
-    "test": "NODE_ENV=test nyc ava --serial",
+    "test": "NODE_ENV=test c8 ava --serial",
     "validate": "npm ls"
   },
   "github": "https://github.com/MatthewCallis/fifo",
@@ -83,7 +88,6 @@
       "src/**/*.js"
     ],
     "require": [
-      "babel-register",
       "./test/helpers/setup-browser-env.js"
     ],
     "concurrency": 8,
