@@ -1,29 +1,49 @@
+/* eslint-disable import/no-extraneous-dependencies */
 /* eslint no-console: 0 */
-const { rollup } = require('rollup');
-const babel = require('rollup-plugin-babel');
-const json = require('rollup-plugin-json');
+import { rollup } from 'rollup';
+import { babel } from '@rollup/plugin-babel';
+import { nodeResolve } from '@rollup/plugin-node-resolve';
+import commonjs from '@rollup/plugin-commonjs';
+import json from '@rollup/plugin-json';
+import cleanup from 'rollup-plugin-cleanup';
 
 rollup({
   input: 'src/fifo.js',
+  external: [
+    'react',
+    'react-dom',
+    'react-datepicker',
+    'react-bootstrap-typeahead',
+    '@babel/runtime',
+  ],
   plugins: [
+    nodeResolve({
+      mainFields: ['module', 'jsnext', 'browser', 'main'],
+    }),
+    commonjs(),
     json({
+      compact: true,
       exclude: ['node_modules/**'],
     }),
     babel({
-      babelrc: false,
-      presets: ['es2015-rollup'],
+      // babelHelpers: 'runtime',
       exclude: 'node_modules/**',
-      plugins: ['external-helpers'],
+      plugins: [],
+    }),
+    cleanup({
+      comments: 'none',
     }),
   ],
-}).then(bundle => (
+}).then((bundle) => (
   bundle.write({
-    format: 'es',
-    name: 'Fifo',
     file: 'lib/fifo.js',
+    format: 'cjs',
+    name: 'fifo',
+    sourcemap: true,
+    exports: 'named',
   })
 )).then(() => {
-  console.log('💝 Fifo Bundle Created!');
+  console.log('Bundle created');
 }).catch((e) => {
-  console.error('💀 Rollup Error:', e);
+  console.log(e);
 });
