@@ -1,3 +1,3 @@
-const browserEnv = require('browser-env');
+import browserEnv from 'browser-env';
 
 browserEnv();
