@@ -1,25 +1,39 @@
-import LocalStorage from './localStorage';
+import LocalStorage from './localStorage.js';
+
+/**
+ * The Fifo configuration object.
+ *
+ * @typedef {object} FifoOptions
+ * @property {string} [namespace] The name of the key in localStorage.
+ * @property {Function} [console] An optional console logging function.
+ * @property {object} [shim] A shim class to act as localStorage in case it is missing.
+ */
+
+/**
+ * A single rule where a subject is compared against a given value on an SDK.
+ *
+ * @typedef {object} FifoData
+ * @property {string[]} keys The collection of keys in the collection.
+ * @property {object} items The actual items in the collection.
+ */
 
 export default class Fifo {
   /**
    * Contructs a new Fifo object.
-   * @param {object} options - The Fifo configuration.
+   *
+   * @param {FifoOptions} options The Fifo configuration.
    */
   constructor(options = {}) {
     this.namespace = options.namespace || 'fifo';
     this.noLS = false;
-
-    if (options.console) {
-      this.console = options.console;
-    } else {
-      this.console = () => {};
-    }
-
-    this.checkLocalStorage(options.shim);
+    this.console = options.console || function console() {};
+    /** @type {FifoData} The localStoage data */
     this.data = {
       keys: [],
       items: {},
     };
+
+    this.checkLocalStorage(options.shim);
 
     const dataFromStorage = JSON.parse(this.LS.getItem(this.namespace));
     if (dataFromStorage && Object.prototype.hasOwnProperty.call(dataFromStorage, 'keys') && Object.prototype.hasOwnProperty.call(dataFromStorage, 'items')) {
@@ -29,9 +43,14 @@ export default class Fifo {
     }
   }
 
+  /**
+   * Checks for the existence of localStorage and shims it should it not exist.
+   * This check has to be wrapped within a try/catch because of a SecurityError: DOM Exception 18 on iOS.
+   *
+   * @param {object} shim A shim class to act as localStorage in case it is missing.
+   */
   checkLocalStorage(shim) {
-    // NOTE: This check has to be wrapped within a try/catch because of a SecurityError: DOM Exception 18 on iOS.
-    /* istanbul ignore next */
+    /* c8 ignore start */
     try {
       if (typeof shim !== 'undefined' || (typeof localStorage !== 'undefined' && localStorage !== null)) {
         this.LS = shim || localStorage;
@@ -46,11 +65,18 @@ export default class Fifo {
       this.noLS = true;
       this.console('warn', 'No localStorage, shimming.');
     }
+    /* c8 ignore stop */
   }
 
   /**
    * Attempts to save the key/value pair to localStorage.
-   * @return {boolean} - Whether or not the save was successful.
+   * Possible Errors:
+   * - 18 for Safari: SecurityError: DOM Exception 18
+   * - 21 for some Safari
+   * - 22 for Chrome and Safari, 1014 for Firefox: QUOTA_EXCEEDED_ERR
+   * - -2147024882 for IE10 Out of Memory
+   *
+   * @returns {boolean} Whether or not the save was successful.
    */
   trySave() {
     if (this.noLS) {
@@ -60,25 +86,20 @@ export default class Fifo {
     try {
       this.LS.setItem(this.namespace, JSON.stringify(this.data));
       return true;
+      /* c8 ignore next 7 */
     } catch (error) {
-      // 18 for Safari: SecurityError: DOM Exception 18
-      // 21 for some Safari
-      // 22 for Chrome and Safari, 1014 for Firefox: QUOTA_EXCEEDED_ERR
-      // -2147024882 for IE10 Out of Memory
-      /* istanbul ignore next */
       if (error.code === 18 || error.code === 21 || error.code === 22 || error.code === 1014 || error.number === -2147024882) {
         return false;
       }
-      /* istanbul ignore next */
       this.console('error', 'Error with localStorage:', error);
-      /* istanbul ignore next */
       return true;
     }
   }
 
   /**
    * Attempts to remove the first item added to make room for the next.
-   * @return The item being removed.
+   *
+   * @returns {object} The item being removed.
    */
   removeFirstIn() {
     const firstIn = this.data.keys.pop();
@@ -92,7 +113,9 @@ export default class Fifo {
 
   /**
    * Save the key/value pair to localStorage.
-   * @return The item being removed.
+   * This is difficult to test without a browser, and difficult in a browser.
+   *
+   * @returns {object[]} The item being removed.
    */
   save() {
     const removed = [];
@@ -100,9 +123,8 @@ export default class Fifo {
       return removed;
     }
 
+    /* c8 ignore next 7 */
     while (!this.trySave()) {
-      // NOTE: Difficult to test without a browser, and difficult in a browser.
-      /* istanbul ignore next */
       if (this.data.keys.length) {
         removed.push(this.removeFirstIn());
       } else {
@@ -115,9 +137,10 @@ export default class Fifo {
 
   /**
    * Set a key/value pair.
-   * @param {string} key - The key to use in the key value pair.
-   * @param value - The value to use in the key value pair.
-   * @return The current instance of Fifo.
+   *
+   * @param {string} key The key to use in the key value pair.
+   * @param {string|number} value The value to use in the key value pair.
+   * @returns {Fifo} The current instance of Fifo.
    */
   set(key, value) {
     this.data.items[key] = value;
@@ -132,8 +155,9 @@ export default class Fifo {
 
   /**
    * Get a value for a given key.
-   * @param {string} [key] - The key to use in the key value pair.
-   * @return The item, or, all items when no key is provided.
+   *
+   * @param {string} [key] The key to use in the key value pair.
+   * @returns {*|[*]} The item, or, all items when no key is provided.
    */
   get(key) {
     if (key) {
@@ -146,7 +170,8 @@ export default class Fifo {
 
   /**
    * All the keys currently being used.
-   * @return {array} The keys.
+   *
+   * @returns {Array} The keys.
    */
   keys() {
     return this.data.keys || [];
@@ -154,8 +179,9 @@ export default class Fifo {
 
   /**
    * Checks for the existence of a key.
+   *
    * @param {string} key - The key to check for.
-   * @return {boolean} Wheter or not the key exist.
+   * @returns {boolean} Wheter or not the key exist.
    */
   has(key) {
     return this.data.keys.indexOf(key) !== -1;
@@ -163,29 +189,27 @@ export default class Fifo {
 
   /**
    * Removes a key/value pair from the collection.
-   * @param victim - The key/value pair to find, can be one of a String, Regular Expression or a Function.
-   * @return The current instance of Fifo.
+   *
+   * @param {string} target - The key/value pair to find, can be one of a String, Regular Expression or a Function.
    */
-  remove(victim) {
-    if (victim == null || !this.data.keys) {
-      return this;
+  remove(target) {
+    if (target == null || !this.data.keys) {
+      return;
     }
 
     const { keys } = this.data;
     keys.forEach((suspect, i) => {
-      if (suspect === victim) {
+      if (suspect === target) {
         this.data.keys.splice(i, 1);
         delete this.data.items[suspect];
       }
     }, this);
 
     this.save();
-    return this;
   }
 
   /**
    * Resets the local data and saves empting out the localStorage to an empty state.
-   * @return The current instance of Fifo.
    */
   empty() {
     this.data = {
@@ -193,6 +217,5 @@ export default class Fifo {
       items: {},
     };
     this.save();
-    return this;
   }
 }
