@@ -1,6 +1,6 @@
-/** @test {Fifo} */
 import test from 'ava';
-import Fifo from '../src/fifo';
+import Fifo from '../src/fifo.js';
+import LocalStorage from '../src/localStorage.js';
 
 // NOTE: Browser test only.
 test.skip('#constructor: will set noLS to true if there is localStorage', (t) => {
@@ -24,7 +24,7 @@ test('#constructor: sets the default console function to be NOOP', (t) => {
 });
 
 test('#constructor: accepts an option to set the console function', (t) => {
-  const newFunction = data => data;
+  const newFunction = (data) => data;
   const collection = new Fifo({ console: newFunction });
   t.is(collection.console, newFunction);
 });
@@ -34,4 +34,13 @@ test('#constructor: inits the data object', (t) => {
   t.is(typeof collection.data, 'object');
   t.true(Array.isArray(collection.data.keys));
   t.is(typeof collection.data.items, 'object');
+});
+
+test('#constructor: can load data from locaStorage values', (t) => {
+  const LS = new LocalStorage();
+  LS.setItem('fifo:test', JSON.stringify({ keys: ['test'], items: { test: 'VALID' } }));
+  const collection = new Fifo({ namespace: 'fifo:test', shim: LS });
+  t.is(typeof collection.data, 'object');
+  t.deepEqual(collection.data.keys, ['test']);
+  t.deepEqual(collection.data.items, { test: 'VALID' });
 });
