@@ -1,8 +1,6 @@
-/** @test {Fifo} */
 import test from 'ava';
-import { iterate } from 'leakage';
-import Fifo from '../src/fifo';
-import LocalStorage from '../src/localStorage';
+import Fifo from '../src/fifo.js';
+import LocalStorage from '../src/localStorage.js';
 
 // With localStorage
 test('#remove: just returns itself on an empty remove', (t) => {
@@ -17,6 +15,14 @@ test('#remove: remove items from storage', (t) => {
   const LS = new LocalStorage();
   const collection = new Fifo({ namespace: 'fifo:test', shim: LS });
   collection.noLS = false;
+
+  collection.remove('foo');
+  t.is(collection.get('foo'), undefined);
+});
+
+test('#remove: remove items from storage (No localStorage)', (t) => {
+  const collection = new Fifo({ namespace: 'fifo:test' });
+  collection.noLS = true;
 
   collection.remove('foo');
   t.is(collection.get('foo'), undefined);
@@ -41,20 +47,11 @@ test('#remove: can remove items using a string', (t) => {
   t.false(collection.has('fifo-regex-key'));
 });
 
-// No localStorage
-test('#remove: just returns itself on an empty remove', (t) => {
+test('#remove: just returns itself on an empty remove (No localStorage)', (t) => {
   const collection = new Fifo({ namespace: 'fifo:test' });
   collection.noLS = true;
 
   t.notThrows(collection.remove);
-});
-
-test('#remove: remove items from storage', (t) => {
-  const collection = new Fifo({ namespace: 'fifo:test' });
-  collection.noLS = true;
-
-  collection.remove('foo');
-  t.is(collection.get('foo'), undefined);
 });
 
 test('#remove: can remove items', (t) => {
@@ -73,21 +70,4 @@ test('#remove: can remove items', (t) => {
   t.false(collection.has('fixed-regex-key'));
   t.true(collection.has('fifo-key'));
   t.false(collection.has('fifo-regex-key'));
-});
-
-test.skip('#remove does not leak', (t) => {
-  t.notThrows(() => {
-    iterate(() => {
-      const collection = new Fifo({ namespace: 'fifo:test' });
-      collection.noLS = true;
-
-      collection.set('fifo-key', 'fifo-regex-value');
-      collection.set('fifo-regex-key', 'fifo-value');
-      collection.set('fixed-key', 'fixed-value');
-      collection.set('fixed-regex-key', 'fixed-regex-value');
-
-      collection.remove('fixed-regex-key');
-      collection.remove('fifo-regex-key');
-    });
-  });
 });
